refactor: use the Div trait to implement div for unsigned integers
