+++ gadgets/src/errors/mod.rs
@@ -16,3 +16,6 @@
 
 pub mod signed_integer;
 pub use self::signed_integer::*;
+
+pub mod unsigned_integer;
+pub use self::unsigned_integer::*;
