+++ gadgets/src/traits/utilities/uint/tests/uint8.rs
@@ -16,6 +16,7 @@
 
 use crate::utilities::{
     alloc::AllocGadget,
+    arithmetic::*,
     boolean::Boolean,
     uint::{UInt, UInt8},
 };
